SHARPNET-684: Split Utils.cs into partials along existing seams
